Crate pango

Rust Pango bindings

Rust bindings and wrappers for Pango, part of gtk-rs-core.

Pango 1.40 is the lowest supported version for the underlying library.

Minimum supported Rust version

Currently, the minimum supported Rust version is 1.83.0.

Documentation

Rust API - Stable

Rust API - Development

C API

GTK Installation instructions

Using

We recommend using crates from crates.io, as demonstrated here.

If you want to track the bleeding edge, use the git dependency instead:

[dependencies]
pango = { git = "https://github.com/gtk-rs/gtk-rs-core.git", package = "pango" }

Avoid mixing versioned and git crates like this:

# This will not compile
[dependencies]
pango = "0.13"
pango = { git = "https://github.com/gtk-rs/gtk-rs-core.git", package = "pango" }

See Also

glib

License

pango is available under the MIT License, please refer to it.

Re-exports

pub use glib;

pub use pango_sys as ffi;

Modules

prelude

Traits and essential types intended for blanket imports.

Structs

Analysis

GLib type: Inline allocated boxed type with stack copy semantics.

AttrClass

AttrColor

GLib type: Boxed type with copy-on-clone semantics.

AttrFloat

GLib type: Boxed type with copy-on-clone semantics.

AttrFontDesc

GLib type: Boxed type with copy-on-clone semantics.

AttrFontFeatures

GLib type: Boxed type with copy-on-clone semantics.

AttrInt

GLib type: Boxed type with copy-on-clone semantics.

AttrIntoIter

AttrIterator

AttrLanguage

GLib type: Boxed type with copy-on-clone semantics.

AttrList

GLib type: Shared boxed type with reference counted clone semantics.

AttrShape

GLib type: Boxed type with copy-on-clone semantics.

AttrSize

GLib type: Boxed type with copy-on-clone semantics.

AttrString

GLib type: Boxed type with copy-on-clone semantics.

Attribute

GLib type: Boxed type with copy-on-clone semantics.

Color

GLib type: Inline allocated boxed type with stack copy semantics.

Context

GLib type: GObject with reference counted clone semantics.

Coverage

GLib type: GObject with reference counted clone semantics.

Font

GLib type: GObject with reference counted clone semantics.

FontDescription

GLib type: Boxed type with copy-on-clone semantics.

FontFace

GLib type: GObject with reference counted clone semantics.

FontFamily

GLib type: GObject with reference counted clone semantics.

FontMap

GLib type: GObject with reference counted clone semantics.

FontMask

FontMetrics

GLib type: Shared boxed type with reference counted clone semantics.

Fontset

GLib type: GObject with reference counted clone semantics.

FontsetSimple

GLib type: GObject with reference counted clone semantics.

GlyphGeometry

GLib type: Inline allocated boxed type with stack copy semantics.

GlyphInfo

GLib type: Inline allocated boxed type with stack copy semantics.

GlyphItem

GLib type: Boxed type with copy-on-clone semantics.

GlyphItemIntoIter

GlyphItemIter

GlyphString

GLib type: Boxed type with copy-on-clone semantics.

HitPosition

The result of LayoutLine::x_to_index.

Item

GLib type: Boxed type with copy-on-clone semantics.

Language

GLib type: Boxed type with copy-on-clone semantics.

Layout

GLib type: GObject with reference counted clone semantics.

LayoutDeserializeFlags

LayoutIter

GLib type: Boxed type with copy-on-clone semantics.

LayoutLine

GLib type: Shared boxed type with reference counted clone semantics.

LayoutSerializeFlags

Matrix

GLib type: Inline allocated boxed type with stack copy semantics.

Rectangle

GLib type: Inline allocated boxed type with stack copy semantics.

Renderer

GLib type: GObject with reference counted clone semantics.

ScriptIntoIter

ScriptIter

ShapeFlags

ShowFlags

TabArray

GLib type: Boxed type with copy-on-clone semantics.

Enums

Alignment

AttrType

BaselineShift

BidiTypeDeprecated

CoverageLevel

Direction

EllipsizeMode

FontScale

Gravity

GravityHint

Overline

RenderPart

Script

Stretch

Style

TabAlign

TextTransform

Underline

Variant

Weight

WrapMode

Constants

ANALYSIS_FLAG_CENTERED_BASELINE

ANALYSIS_FLAG_IS_ELLIPSIS

ANALYSIS_FLAG_NEED_HYPHEN

ATTR_INDEX_FROM_TEXT_BEGINNING

ATTR_INDEX_TO_TEXT_END

GLYPH_EMPTY

GLYPH_INVALID_INPUT

GLYPH_UNKNOWN_FLAG

SCALE

SCALE_LARGE

The scale factor for one magnification step (1.2).

SCALE_MEDIUM

The scale factor for normal size (1.0).

SCALE_SMALL

The scale factor for one shrinking step (1 / 1.2).

SCALE_XX_LARGE

The scale factor for three magnification steps (1.2 * 1.2 * 1.2).

SCALE_XX_SMALL

The scale factor for three shrinking steps (1 / (1.2 * 1.2 * 1.2)).

SCALE_X_LARGE

The scale factor for two magnification steps (1.2 * 1.2).

SCALE_X_SMALL

The scale factor for two shrinking steps (1 / (1.2 * 1.2)).

Traits

IsAttribute

Functions

extents_to_pixels

find_base_dir

find_paragraph_boundary

is_zero_width

itemize

itemize_with_base_dir

markup_parser_finish

markup_parser_new

parse_markup

parse_stretch

parse_style

parse_variant

parse_weight

quantize_line_geometry

reorder_items

shape

shape_full

shape_with_flags

unichar_direction

units_from_double

units_to_double

version

version_check

version_string

Type Aliases

Glyph

GlyphUnit

LayoutRun