Function pango::extents_to_pixels

pub fn extents_to_pixels(
    inclusive: Option<&mut Rectangle>,
    nearest: Option<&mut Rectangle>
)

Converts extents from Pango units to device units.

The conversion is done by dividing by the PANGO_SCALE factor and performing rounding.

The @inclusive rectangle is converted by flooring the x/y coordinates and extending width/height, such that the final rectangle completely includes the original rectangle.

The @nearest rectangle is converted by rounding the coordinates of the rectangle to the nearest device unit (pixel).

The rule to which argument to use is: if you want the resulting device-space rectangle to completely contain the original rectangle, pass it in as @inclusive. If you want two touching-but-not-overlapping rectangles stay touching-but-not-overlapping after rounding to device units, pass them in as @nearest.

inclusive

rectangle to round to pixels inclusively

nearest

rectangle to round to nearest pixels