pango

Function find_base_dir

pub fn find_base_dir(text: &str) -> Direction

Searches a string the first character that has a strong direction, according to the Unicode bidirectional algorithm.

text

the text to process. Must be valid UTF-8

length

length of @text in bytes (may be -1 if @text is nul-terminated)

Returns

The direction corresponding to the first strong character. If no such character is found, then Direction::Neutral is returned.