pango

Function is_zero_width

pub fn is_zero_width(ch: char) -> bool

Checks if a character that should not be normally rendered.

This includes all Unicode characters with “ZERO WIDTH” in their name, as well as bidi formatting characters, and a few other ones.

This is totally different from unichar_iszerowidth() and is at best misnamed.

ch

a Unicode character

Returns

true if @ch is a zero-width character, false otherwise