pango

Function markup_parser_finish

pub fn markup_parser_finish(
    context: &MarkupParseContext,
) -> Result<(AttrList, GString, char), Error>

Finishes parsing markup.

After feeding a Pango markup parser some data with glib::MarkupParseContext::parse(), use this function to get the list of attributes and text out of the markup. This function will not free @context, use GLib::MarkupParseContext::free() to do so.

context

A valid parse context that was returned from markup_parser_new()

Returns

false if @error is set, otherwise true

attr_list

address of return location for a AttrList

text

address of return location for text with tags stripped

accel_char

address of return location for accelerator char