Function pango::parse_markup

pub fn parse_markup(
    markup_text: &str, 
    accel_marker: char
) -> Result<(AttrList, GString, char), Error>

Parses marked-up text to create a plain-text string and an attribute list.

See the Pango Markup docs for details about the supported markup.

If accel_marker is nonzero, the given character will mark the character following it as an accelerator. For example, accel_marker might be an ampersand or underscore. All characters marked as an accelerator will receive a Underline::Low attribute, and the first character so marked will be returned in accel_char. Two accel_marker characters following each other produce a single literal accel_marker character.

To parse a stream of pango markup incrementally, use markup_parser_new().

If any error happens, none of the output arguments are touched except for error.

markup_text

markup to parse (see the Pango Markup docs)

length

length of markup_text, or -1 if nul-terminated

accel_marker

character that precedes an accelerator, or 0 for none

Returns

false if error is set, otherwise true

attr_list

address of return location for a AttrList

text

address of return location for text with tags stripped

accel_char

address of return location for accelerator char