Function pango::quantize_line_geometry

pub fn quantize_line_geometry(thickness: &mut i32, position: &mut i32)

Quantizes the thickness and position of a line, typically an underline or strikethrough, to whole device pixels, that is integer multiples of PANGO_SCALE. The purpose of this function is to avoid such lines looking blurry.

Care is taken to make sure thickness is at least one pixel when this function returns, but returned position may become zero as a result of rounding.

thickness

pointer to the thickness of a line, in Pango units

position

corresponding position