Function pango::reorder_items

pub fn reorder_items(logical_items: &[&Item]) -> Vec<Item>

From a list of items in logical order and the associated directional levels, produce a list in visual order. The original list is unmodified.

logical_items

a [crate::glib::List] of crate::Item in logical order.

Returns

a [crate::glib::List] of crate::Item structures in visual order.

(Please open a bug if you use this function. It is not a particularly convenient interface, and the code is duplicated elsewhere in Pango for that reason.)