Function pango::shape_with_flags

pub fn shape_with_flags(
    item_text: &str, 
    paragraph_text: Option<&str>, 
    analysis: &Analysis, 
    glyphs: &mut GlyphString, 
    flags: ShapeFlags
)

This is supported on crate feature v1_44 only.

Given a segment of text and the corresponding crate::Analysis structure returned from pango_itemize(), convert the characters into glyphs. You may also pass in only a substring of the item from pango_itemize().

This is similar to pango_shape_full(), except it also takes flags that can influence the shaping process.

item_text

valid UTF-8 text to shape

item_length

the length (in bytes) of item_text. -1 means nul-terminated text.

paragraph_text

text of the paragraph (see details). May be None.

paragraph_length

the length (in bytes) of paragraph_text. -1 means nul-terminated text.

analysis

crate::Analysis structure from pango_itemize()

glyphs

glyph string in which to store results

flags

flags influencing the shaping process