pango

Function version_check

pub fn version_check(
    required_major: i32,
    required_minor: i32,
    required_micro: i32,
) -> Option<GString>

Checks that the Pango library in use is compatible with the given version.

Generally you would pass in the constants PANGO_VERSION_MAJOR, PANGO_VERSION_MINOR, PANGO_VERSION_MICRO as the three arguments to this function; that produces a check that the library in use at run-time is compatible with the version of Pango the application or module was compiled against.

Compatibility is defined by two things: first the version of the running library is newer than the version @required_major.required_minor.@required_micro. Second the running library must be binary compatible with the version @required_major.required_minor.@required_micro (same major version.)

For compile-time version checking use PANGO_VERSION_CHECK().

required_major

the required major version

required_minor

the required minor version

required_micro

the required major version

Returns

None if the Pango library is compatible with the given version, or a string describing the version mismatch. The returned string is owned by Pango and should not be modified or freed.