pango

Function version_string

pub fn version_string() -> GString

Returns the version of Pango available at run-time.

This is similar to the macro PANGO_VERSION_STRING except that the macro returns the version available at compile-time.

Returns

A string containing the version of Pango library available at run time. The returned string is owned by Pango and should not be modified or freed.