Trait pango::prelude::FontFaceExt

pub trait FontFaceExt: 'static {
    fn describe(&self) -> Option<FontDescription>;

    fn get_face_name(&self) -> Option<GString>;

    fn get_family(&self) -> Option<FontFamily>;

    fn is_synthesized(&self) -> bool;

    fn list_sizes(&self) -> Vec<i32>;
}

Trait containing all FontFace methods.

Implementors

FontFace

Required methods

fn describe(&self) -> Option<FontDescription>

Returns the family, style, variant, weight and stretch of a FontFace. The size field of the resulting font description will be unset.

Returns

a newly-created FontDescription structure holding the description of the face. Use FontDescription::free to free the result.

fn get_face_name(&self) -> Option<GString>

Gets a name representing the style of this face among the different faces in the FontFamily for the face. This name is unique among all faces in the family and is suitable for displaying to users.

Returns

the face name for the face. This string is owned by the face object and must not be modified or freed.

fn get_family(&self) -> Option<FontFamily>

This is supported on crate feature v1_46 only.

Gets the FontFamily that self belongs to.

Feature: v1_46

Returns

the FontFamily

fn is_synthesized(&self) -> bool

Returns whether a FontFace is synthesized by the underlying font rendering engine from another face, perhaps by shearing, emboldening, or lightening it.

Returns

whether self is synthesized.

fn list_sizes(&self) -> Vec<i32>

List the available sizes for a font. This is only applicable to bitmap fonts. For scalable fonts, stores None at the location pointed to by sizes and 0 at the location pointed to by n_sizes. The sizes returned are in Pango units and are sorted in ascending order.

sizes

location to store a pointer to an array of int. This array should be freed with g_free.

n_sizes

location to store the number of elements in sizes

Loading content...