Struct pango::AttrLanguage

pub struct AttrLanguage { /* private fields */ }

GLib type: Boxed type with copy-on-clone semantics.

Implementations

impl AttrLanguage

pub fn as_ptr(&self) -> *mut PangoAttrLanguage

Return the inner pointer to the underlying C value.

pub unsafe fn from_glib_ptr_borrow(ptr: &*mut PangoAttrLanguage) -> &Self

Borrows the underlying C value.

pub unsafe fn from_glib_ptr_borrow_mut( ptr: &mut *mut PangoAttrLanguage, ) -> &mut Self

Borrows the underlying C value mutably.

impl AttrLanguage

pub fn new(language: &Language) -> Self

Create a new language tag attribute.

language

language tag

Returns

the newly allocated Attribute, which should be freed with Pango::Attribute::destroy()

pub fn value(&self) -> Language