Struct pango::Color

#[repr(transparent)]
pub struct Color { /* private fields */ }

The Color structure is used to represent a color in an uncalibrated RGB color-space.

Implementations

impl Color

pub fn to_str(&self) -> GString

Returns a textual specification of @self.

The string is in the hexadecimal form #rrrrggggbbbb, where r, g and b are hex digits representing the red, green, and blue components respectively.

Returns

a newly-allocated text string that must be freed with g_free().

impl Color

pub fn parse(spec: &str) -> Result<Self, BoolError>

Fill in the fields of a color from a string specification.

The string can either one of a large set of standard names. (Taken from the CSS Color specification, or it can be a value in the form #rgb, #rrggbb, #rrrgggbbb or #rrrrggggbbbb, where r, g and b are hex digits of the red, green, and blue components of the color, respectively. (White in the four forms is #fff, #ffffff, #fffffffff and #ffffffffffff.)

spec

a string specifying the new color

Returns

true if parsing of the specifier succeeded, otherwise false

pub fn parse_with_alpha(spec: &str) -> Result<(Self, u16), BoolError>

Available on crate feature v1_46 only.

Fill in the fields of a color from a string specification.

The string can either one of a large set of standard names. (Taken from the CSS Color specification, or it can be a hexadecimal value in the form #rgb, #rrggbb, #rrrgggbbb or #rrrrggggbbbb where r, g and b are hex digits of the red, green, and blue components of the color, respectively. (White in the four forms is #fff, #ffffff, #fffffffff and #ffffffffffff.)

Additionally, parse strings of the form #rgba, #rrggbbaa, #rrrrggggbbbbaaaa, if @alpha is not None, and set @alpha to the value specified by the hex digits for a. If no alpha component is found in @spec, @alpha is set to 0xffff (for a solid color).

spec

a string specifying the new color

Returns

true if parsing of the specifier succeeded, otherwise false

alpha

return location for alpha

pub fn red(&self) -> u16

pub fn green(&self) -> u16

pub fn blue(&self) -> u16