Struct pango::Context

pub struct Context(_);

Implementations

impl Context

pub fn new() -> Context

pub fn changed(&self)

Forces a change in the context, which will cause any Layout using this context to re-layout.

This function is only useful when implementing a new backend for Pango, something applications won't do. Backends should call this function if they have attached extra data to the context and such data is changed.

pub fn get_base_dir(&self) -> Direction

Retrieves the base direction for the context. See Context::set_base_dir.

Returns

the base direction for the context.

pub fn get_base_gravity(&self) -> Gravity

Retrieves the base gravity for the context. See Context::set_base_gravity.

Returns

the base gravity for the context.

pub fn get_font_description(&self) -> Option<FontDescription>

Retrieve the default font description for the context.

Returns

a pointer to the context's default font description. This value must not be modified or freed.

pub fn get_font_map(&self) -> Option<FontMap>

Gets the FontMap used to look up fonts for this context.

Returns

the font map for the Context. This value is owned by Pango and should not be unreferenced.

pub fn get_gravity(&self) -> Gravity

Retrieves the gravity for the context. This is similar to Context::get_base_gravity, except for when the base gravity is Gravity::Auto for which Gravity::get_for_matrix is used to return the gravity from the current context matrix.

Returns

the resolved gravity for the context.

pub fn get_gravity_hint(&self) -> GravityHint

Retrieves the gravity hint for the context. See Context::set_gravity_hint for details.

Returns

the gravity hint for the context.

pub fn get_language(&self) -> Option<Language>

Retrieves the global language tag for the context.

Returns

the global language tag.

pub fn get_matrix(&self) -> Option<Matrix>

pub fn get_metrics(
&self,
desc: Option<&FontDescription>,
language: Option<&Language>
) -> Option<FontMetrics>

Get overall metric information for a particular font description. Since the metrics may be substantially different for different scripts, a language tag can be provided to indicate that the metrics should be retrieved that correspond to the script(s) used by that language.

The FontDescription is interpreted in the same way as by pango_itemize, and the family name may be a comma separated list of figures. If characters from multiple of these families would be used to render the string, then the returned fonts would be a composite of the metrics for the fonts loaded for the individual families.

desc

a FontDescription structure. None means that the font description from the context will be used.

language

language tag used to determine which script to get the metrics for. None means that the language tag from the context will be used. If no language tag is set on the context, metrics for the default language (as determined by Language::get_default) will be returned.

Returns

a FontMetrics object. The caller must call FontMetrics::unref when finished using the object.

pub fn get_round_glyph_positions(&self) -> bool

This is supported on crate feature v1_44 only.

Returns whether font rendering with this context should round glyph positions and widths.

Feature: v1_44

pub fn get_serial(&self) -> u32

Returns the current serial number of self. The serial number is initialized to an small number larger than zero when a new context is created and is increased whenever the context is changed using any of the setter functions, or the FontMap it uses to find fonts has changed. The serial may wrap, but will never have the value 0. Since it can wrap, never compare it with "less than", always use "not equals".

This can be used to automatically detect changes to a Context, and is only useful when implementing objects that need update when their Context changes, like Layout.

Returns

The current serial number of self.

pub fn list_families(&self) -> Vec<FontFamily>

List all families for a context.

families

location to store a pointer to an array of FontFamily *. This array should be freed with g_free.

n_families

location to store the number of elements in descs

pub fn load_font(&self, desc: &FontDescription) -> Option<Font>

Loads the font in one of the fontmaps in the context that is the closest match for desc.

desc

a FontDescription describing the font to load

Returns

the newly allocated Font that was loaded, or None if no font matched.

pub fn load_fontset(
&self,
desc: &FontDescription,
language: &Language
) -> Option<Fontset>

Load a set of fonts in the context that can be used to render a font matching desc.

desc

a FontDescription describing the fonts to load

language

a Language the fonts will be used for

Returns

the newly allocated Fontset loaded, or None if no font matched.

pub fn set_base_dir(&self, direction: Direction)

Sets the base direction for the context.

The base direction is used in applying the Unicode bidirectional algorithm; if the direction is Direction::Ltr or Direction::Rtl, then the value will be used as the paragraph direction in the Unicode bidirectional algorithm. A value of Direction::WeakLtr or Direction::WeakRtl is used only for paragraphs that do not contain any strong characters themselves.

direction

the new base direction

pub fn set_base_gravity(&self, gravity: Gravity)

Sets the base gravity for the context.

The base gravity is used in laying vertical text out.

gravity

the new base gravity

pub fn set_font_description(&self, desc: &FontDescription)

Set the default font description for the context

desc

the new pango font description

pub fn set_font_map<P: IsA<FontMap>>(&self, font_map: &P)

Sets the font map to be searched when fonts are looked-up in this context. This is only for internal use by Pango backends, a Context obtained via one of the recommended methods should already have a suitable font map.

font_map

the FontMap to set.

pub fn set_gravity_hint(&self, hint: GravityHint)

Sets the gravity hint for the context.

The gravity hint is used in laying vertical text out, and is only relevant if gravity of the context as returned by Context::get_gravity is set Gravity::East or Gravity::West.

hint

the new gravity hint

pub fn set_language(&self, language: &Language)

Sets the global language tag for the context. The default language for the locale of the running process can be found using Language::get_default.

language

the new language tag.

pub fn set_matrix(&self, matrix: Option<&Matrix>)

pub fn set_round_glyph_positions(&self, round_positions: bool)

This is supported on crate feature v1_44 only.

Sets whether font rendering with this context should round glyph positions and widths to integral positions, in device units.

This is useful when the renderer can't handle subpixel positioning of glyphs.

The default value is to round glyph positions, to remain compatible with previous Pango behavior.

Feature: v1_44

round_positions

whether to round glyph positions