Struct pango::FontDescription

pub struct FontDescription(_);

The crate::FontDescription structure represents the description of an ideal font. These structures are used both to list what fonts are available on the system and also for specifying the characteristics of a font to load.

Implementations

impl FontDescription

pub fn new() -> FontDescription

Creates a new font description structure with all fields unset.

Returns

the newly allocated crate::FontDescription, which should be freed using [Self::free()].

pub fn better_match(
&self,
old_match: Option<&FontDescription>,
new_match: &FontDescription
) -> bool

Determines if the style attributes of new_match are a closer match for self than those of old_match are, or if old_match is None, determines if new_match is a match at all. Approximate matching is done for weight and style; other style attributes must match exactly. Style attributes are all attributes other than family and size-related attributes. Approximate matching for style considers PANGO_STYLE_OBLIQUE and PANGO_STYLE_ITALIC as matches, but not as good a match as when the styles are equal.

Note that old_match must match self.

old_match

a crate::FontDescription, or None

new_match

a crate::FontDescription

Returns

true if new_match is a better match

pub fn family(&self) -> Option<GString>

pub fn gravity(&self) -> Gravity

pub fn set_fields(&self) -> FontMask

pub fn size(&self) -> i32

pub fn is_size_absolute(&self) -> bool

pub fn stretch(&self) -> Stretch

pub fn style(&self) -> Style

pub fn variant(&self) -> Variant

pub fn variations(&self) -> Option<GString>

This is supported on crate feature v1_42 only.

pub fn weight(&self) -> Weight

pub fn merge(
&mut self,
desc_to_merge: Option<&FontDescription>,
replace_existing: bool
)

Merges the fields that are set in desc_to_merge into the fields in self. If replace_existing is false, only fields in self that are not already set are affected. If true, then fields that are already set will be replaced as well.

If desc_to_merge is None, this function performs nothing.

desc_to_merge

the crate::FontDescription to merge from, or None

replace_existing

if true, replace fields in self with the corresponding values from desc_to_merge, even if they are already exist.

pub fn set_absolute_size(&mut self, size: f64)

Sets the size field of a font description, in device units. This is mutually exclusive with Self::set_size() which sets the font size in points.

size

the new size, in Pango units. There are PANGO_SCALE Pango units in one device unit. For an output backend where a device unit is a pixel, a size value of 10 * PANGO_SCALE gives a 10 pixel font.

pub fn set_family(&mut self, family: &str)

Sets the family name field of a font description. The family name represents a family of related font styles, and will resolve to a particular crate::FontFamily. In some uses of crate::FontDescription, it is also possible to use a comma separated list of family names for this field.

family

a string representing the family name.

pub fn set_gravity(&mut self, gravity: Gravity)

Sets the gravity field of a font description. The gravity field specifies how the glyphs should be rotated. If gravity is crate::Gravity::Auto, this actually unsets the gravity mask on the font description.

This function is seldom useful to the user. Gravity should normally be set on a crate::Context.

gravity

the gravity for the font description.

pub fn set_size(&mut self, size: i32)

Sets the size field of a font description in fractional points. This is mutually exclusive with Self::set_absolute_size().

size

the size of the font in points, scaled by PANGO_SCALE. (That is, a size value of 10 * PANGO_SCALE is a 10 point font. The conversion factor between points and device units depends on system configuration and the output device. For screen display, a logical DPI of 96 is common, in which case a 10 point font corresponds to a 10 * (96 / 72) = 13.3 pixel font. Use Self::set_absolute_size() if you need a particular size in device units.

pub fn set_stretch(&mut self, stretch: Stretch)

Sets the stretch field of a font description. The stretch field specifies how narrow or wide the font should be.

stretch

the stretch for the font description

pub fn set_style(&mut self, style: Style)

Sets the style field of a crate::FontDescription. The crate::Style enumeration describes whether the font is slanted and the manner in which it is slanted; it can be either crate::Style::Normal, crate::Style::Italic, or crate::Style::Oblique. Most fonts will either have a italic style or an oblique style, but not both, and font matching in Pango will match italic specifications with oblique fonts and vice-versa if an exact match is not found.

style

the style for the font description

pub fn set_variant(&mut self, variant: Variant)

Sets the variant field of a font description. The crate::Variant can either be crate::Variant::Normal or crate::Variant::SmallCaps.

variant

the variant type for the font description.

pub fn set_variations(&mut self, variations: &str)

This is supported on crate feature v1_42 only.

Sets the variations field of a font description. OpenType font variations allow to select a font instance by specifying values for a number of axes, such as width or weight.

The format of the variations string is AXIS1=VALUE,AXIS2=VALUE…, with each AXIS a 4 character tag that identifies a font axis, and each VALUE a floating point number. Unknown axes are ignored, and values are clamped to their allowed range.

Pango does not currently have a way to find supported axes of a font. Both harfbuzz or freetype have API for this.

variations

a string representing the variations

pub fn set_variations_static(&mut self, variations: &str)

This is supported on crate feature v1_42 only.

Like Self::set_variations(), except that no copy of variations is made. The caller must make sure that the string passed in stays around until self has been freed or the name is set again. This function can be used if variations is a static string such as a C string literal, or if self is only needed temporarily.

variations

a string representing the variations

pub fn set_weight(&mut self, weight: Weight)

Sets the weight field of a font description. The weight field specifies how bold or light the font should be. In addition to the values of the crate::Weight enumeration, other intermediate numeric values are possible.

weight

the weight for the font description.

pub fn to_filename(&self) -> Option<GString>

Creates a filename representation of a font description. The filename is identical to the result from calling [Self::to_string()], but with underscores instead of characters that are untypical in filenames, and in lower case only.

Returns

a new string that must be freed with g_free().

pub fn to_str(&self) -> GString

pub fn unset_fields(&mut self, to_unset: FontMask)

Unsets some of the fields in a crate::FontDescription. The unset fields will get back to their default values.

to_unset

bitmask of fields in the self to unset.

pub fn from_string(str: &str) -> FontDescription

Creates a new font description from a string representation in the form

“[FAMILY-LIST] [STYLE-OPTIONS] [SIZE] [VARIATIONS]”,

where FAMILY-LIST is a comma-separated list of families optionally terminated by a comma, STYLE_OPTIONS is a whitespace-separated list of words where each word describes one of style, variant, weight, stretch, or gravity, and SIZE is a decimal number (size in points) or optionally followed by the unit modifier “px” for absolute size. VARIATIONS is a comma-separated list of font variation specifications of the form “@axis=value” (the = sign is optional).

The following words are understood as styles: “Normal”, “Roman”, “Oblique”, “Italic”.

The following words are understood as variants: “Small-Caps”.

The following words are understood as weights: “Thin”, “Ultra-Light”, “Extra-Light”, “Light”, “Semi-Light”, “Demi-Light”, “Book”, “Regular”, “Medium”, “Semi-Bold”, “Demi-Bold”, “Bold”, “Ultra-Bold”, “Extra-Bold”, “Heavy”, “Black”, “Ultra-Black”, “Extra-Black”.

The following words are understood as stretch values: “Ultra-Condensed”, “Extra-Condensed”, “Condensed”, “Semi-Condensed”, “Semi-Expanded”, “Expanded”, “Extra-Expanded”, “Ultra-Expanded”.

The following words are understood as gravity values: “Not-Rotated”, “South”, “Upside-Down”, “North”, “Rotated-Left”, “East”, “Rotated-Right”, “West”.

Any one of the options may be absent. If FAMILY-LIST is absent, then the family_name field of the resulting font description will be initialized to None. If STYLE-OPTIONS is missing, then all style options will be set to the default values. If SIZE is missing, the size in the resulting font description will be set to 0.

A typical example:

“Cantarell Italic Light 15 @wght=200”

str

string representation of a font description.

Returns

a new crate::FontDescription.