Struct pango::LayoutIter

pub struct LayoutIter(_);

A LayoutIter structure can be used to iterate over the visual extents of a Layout.

The LayoutIter structure is opaque, and has no user-visible fields.

Implementations

impl LayoutIter

pub fn at_last_line(&mut self) -> bool

Determines whether self is on the last line of the layout.

Returns

true if self is on the last line.

pub fn baseline(&mut self) -> i32

Gets the Y position of the current line’s baseline, in layout coordinates (origin at top left of the entire layout).

Returns

baseline of current line.

pub fn char_extents(&mut self) -> Rectangle

Gets the extents of the current character, in layout coordinates (origin is the top left of the entire layout). Only logical extents can sensibly be obtained for characters; ink extents make sense only down to the level of clusters.

Returns

logical_rect

rectangle to fill with logical extents

pub fn cluster_extents(&mut self) -> (Rectangle, Rectangle)

Gets the extents of the current cluster, in layout coordinates (origin is the top left of the entire layout).

Returns

ink_rect

rectangle to fill with ink extents, or None

logical_rect

rectangle to fill with logical extents, or None

pub fn index(&mut self) -> i32

Gets the current byte index. Note that iterating forward by char moves in visual order, not logical order, so indexes may not be sequential. Also, the index may be equal to the length of the text in the layout, if on the None run (see run()).

Returns

current byte index.

pub fn layout(&mut self) -> Option<Layout>

Gets the layout associated with a LayoutIter.

Returns

the layout associated with self.

pub fn layout_extents(&mut self) -> (Rectangle, Rectangle)

Obtains the extents of the Layout being iterated over. ink_rect or logical_rect can be None if you aren’t interested in them.

Returns

ink_rect

rectangle to fill with ink extents, or None

logical_rect

rectangle to fill with logical extents, or None

pub fn line(&mut self) -> Option<LayoutLine>

Gets the current line.

Use the faster line_readonly() if you do not plan to modify the contents of the line (glyphs, glyph widths, etc.).

Returns

the current line.

pub fn line_extents(&mut self) -> (Rectangle, Rectangle)

Obtains the extents of the current line. ink_rect or logical_rect can be None if you aren’t interested in them. Extents are in layout coordinates (origin is the top-left corner of the entire Layout). Thus the extents returned by this function will be the same width/height but not at the same x/y as the extents returned from LayoutLine::extents().

Returns

ink_rect

rectangle to fill with ink extents, or None

logical_rect

rectangle to fill with logical extents, or None

pub fn line_readonly(&mut self) -> Option<LayoutLine>

Gets the current line for read-only access.

This is a faster alternative to line(), but the user is not expected to modify the contents of the line (glyphs, glyph widths, etc.).

Returns

the current line, that should not be modified.

pub fn line_yrange(&mut self) -> (i32, i32)

Divides the vertical space in the Layout being iterated over between the lines in the layout, and returns the space belonging to the current line. A line’s range includes the line’s logical extents, plus half of the spacing above and below the line, if Layout::set_spacing() has been called to set layout spacing. The Y positions are in layout coordinates (origin at top left of the entire layout).

Note: Since 1.44, Pango uses line heights for placing lines, and there may be gaps between the ranges returned by this function.

Returns

y0_

start of line, or None

y1_

end of line, or None

pub fn run(&mut self) -> Option<LayoutRun>

Gets the current run. When iterating by run, at the end of each line, there’s a position with a None run, so this function can return None. The None run at the end of each line ensures that all lines have at least one run, even lines consisting of only a newline.

Use the faster run_readonly() if you do not plan to modify the contents of the run (glyphs, glyph widths, etc.).

Returns

the current run.

pub fn run_extents(&mut self) -> (Rectangle, Rectangle)

Gets the extents of the current run in layout coordinates (origin is the top left of the entire layout).

Returns

ink_rect

rectangle to fill with ink extents, or None

logical_rect

rectangle to fill with logical extents, or None

pub fn run_readonly(&mut self) -> Option<LayoutRun>

Gets the current run. When iterating by run, at the end of each line, there’s a position with a None run, so this function can return None. The None run at the end of each line ensures that all lines have at least one run, even lines consisting of only a newline.

This is a faster alternative to run(), but the user is not expected to modify the contents of the run (glyphs, glyph widths, etc.).

Returns

the current run, that should not be modified.

pub fn next_char(&mut self) -> bool

Moves self forward to the next character in visual order. If self was already at the end of the layout, returns false.

Returns

whether motion was possible.

pub fn next_cluster(&mut self) -> bool

Moves self forward to the next cluster in visual order. If self was already at the end of the layout, returns false.

Returns

whether motion was possible.

pub fn next_line(&mut self) -> bool

Moves self forward to the start of the next line. If self is already on the last line, returns false.

Returns

whether motion was possible.

pub fn next_run(&mut self) -> bool

Moves self forward to the next run in visual order. If self was already at the end of the layout, returns false.

Returns

whether motion was possible.