Trait pango::ContextExt

pub trait ContextExt: 'static {
    fn changed(&self);

    fn get_base_dir(&self) -> Direction;

    fn get_base_gravity(&self) -> Gravity;

    fn get_font_description(&self) -> Option<FontDescription>;

    fn get_font_map(&self) -> Option<FontMap>;

    fn get_gravity(&self) -> Gravity;

    fn get_gravity_hint(&self) -> GravityHint;

    fn get_language(&self) -> Option<Language>;

    fn get_matrix(&self) -> Option<Matrix>;

    fn get_metrics<'a, 'b, P: Into<Option<&'a FontDescription>>, Q: Into<Option<&'b Language>>>(
        &self, 
        desc: P, 
        language: Q
    ) -> Option<FontMetrics>;

    fn get_serial(&self) -> u32;

    fn list_families(&self) -> Vec<FontFamily>;

    fn load_font(&self, desc: &FontDescription) -> Option<Font>;

    fn load_fontset(
        &self, 
        desc: &FontDescription, 
        language: &Language
    ) -> Option<Fontset>;

    fn set_base_dir(&self, direction: Direction);

    fn set_base_gravity(&self, gravity: Gravity);

    fn set_font_description(&self, desc: &FontDescription);

    fn set_font_map<P: IsA<FontMap>>(&self, font_map: &P);

    fn set_gravity_hint(&self, hint: GravityHint);

    fn set_language(&self, language: &Language);

    fn set_matrix<'a, P: Into<Option<&'a Matrix>>>(&self, matrix: P);
}

Trait containing all Context methods.

Implementors

Context

Required methods

fn changed(&self)

Forces a change in the context, which will cause any Layout using this context to re-layout.

This function is only useful when implementing a new backend for Pango, something applications won't do. Backends should call this function if they have attached extra data to the context and such data is changed.

fn get_base_dir(&self) -> Direction

Retrieves the base direction for the context. See ContextExt::set_base_dir.

Returns

the base direction for the context.

fn get_base_gravity(&self) -> Gravity

Retrieves the base gravity for the context. See ContextExt::set_base_gravity.

Returns

the base gravity for the context.

fn get_font_description(&self) -> Option<FontDescription>

Retrieve the default font description for the context.

Returns

a pointer to the context's default font description. This value must not be modified or freed.

fn get_font_map(&self) -> Option<FontMap>

Gets the FontMap used to look up fonts for this context.

Returns

the font map for the Context. This value is owned by Pango and should not be unreferenced.

fn get_gravity(&self) -> Gravity

Retrieves the gravity for the context. This is similar to ContextExt::get_base_gravity, except for when the base gravity is Gravity::Auto for which Gravity::get_for_matrix is used to return the gravity from the current context matrix.

Returns

the resolved gravity for the context.

fn get_gravity_hint(&self) -> GravityHint

Retrieves the gravity hint for the context. See ContextExt::set_gravity_hint for details.

Returns

the gravity hint for the context.

fn get_language(&self) -> Option<Language>

Retrieves the global language tag for the context.

Returns

the global language tag.

fn get_matrix(&self) -> Option<Matrix>

Gets the transformation matrix that will be applied when rendering with this context. See ContextExt::set_matrix.

Returns

the matrix, or None if no matrix has been set (which is the same as the identity matrix). The returned matrix is owned by Pango and must not be modified or freed.

fn get_metrics<'a, 'b, P: Into<Option<&'a FontDescription>>, Q: Into<Option<&'b Language>>>(
&self,
desc: P,
language: Q
) -> Option<FontMetrics>

Get overall metric information for a particular font description. Since the metrics may be substantially different for different scripts, a language tag can be provided to indicate that the metrics should be retrieved that correspond to the script(s) used by that language.

The FontDescription is interpreted in the same way as by pango_itemize, and the family name may be a comma separated list of figures. If characters from multiple of these families would be used to render the string, then the returned fonts would be a composite of the metrics for the fonts loaded for the individual families.

desc

a FontDescription structure. None means that the font description from the context will be used.

language

language tag used to determine which script to get the metrics for. None means that the language tag from the context will be used. If no language tag is set on the context, metrics for the default language (as determined by Language::get_default) will be returned.

Returns

a FontMetrics object. The caller must call FontMetrics::unref when finished using the object.

fn get_serial(&self) -> u32

Returns the current serial number of self. The serial number is initialized to an small number larger than zero when a new context is created and is increased whenever the context is changed using any of the setter functions, or the FontMap it uses to find fonts has changed. The serial may wrap, but will never have the value 0. Since it can wrap, never compare it with "less than", always use "not equals".

This can be used to automatically detect changes to a Context, and is only useful when implementing objects that need update when their Context changes, like Layout.

Returns

The current serial number of self.

fn list_families(&self) -> Vec<FontFamily>

List all families for a context.

families

location to store a pointer to an array of FontFamily *. This array should be freed with g_free.

n_families

location to store the number of elements in descs

fn load_font(&self, desc: &FontDescription) -> Option<Font>

Loads the font in one of the fontmaps in the context that is the closest match for desc.

desc

a FontDescription describing the font to load

Returns

the newly allocated Font that was loaded, or None if no font matched.

fn load_fontset(
&self,
desc: &FontDescription,
language: &Language
) -> Option<Fontset>

Load a set of fonts in the context that can be used to render a font matching desc.

desc

a FontDescription describing the fonts to load

language

a Language the fonts will be used for

Returns

the newly allocated Fontset loaded, or None if no font matched.

fn set_base_dir(&self, direction: Direction)

Sets the base direction for the context.

The base direction is used in applying the Unicode bidirectional algorithm; if the direction is Direction::Ltr or Direction::Rtl, then the value will be used as the paragraph direction in the Unicode bidirectional algorithm. A value of Direction::WeakLtr or Direction::WeakRtl is used only for paragraphs that do not contain any strong characters themselves.

direction

the new base direction

fn set_base_gravity(&self, gravity: Gravity)

Sets the base gravity for the context.

The base gravity is used in laying vertical text out.

gravity

the new base gravity

fn set_font_description(&self, desc: &FontDescription)

Set the default font description for the context

desc

the new pango font description

fn set_font_map<P: IsA<FontMap>>(&self, font_map: &P)

Sets the font map to be searched when fonts are looked-up in this context. This is only for internal use by Pango backends, a Context obtained via one of the recommended methods should already have a suitable font map.

font_map

the FontMap to set.

fn set_gravity_hint(&self, hint: GravityHint)

Sets the gravity hint for the context.

The gravity hint is used in laying vertical text out, and is only relevant if gravity of the context as returned by ContextExt::get_gravity is set Gravity::East or Gravity::West.

hint

the new gravity hint

fn set_language(&self, language: &Language)

Sets the global language tag for the context. The default language for the locale of the running process can be found using Language::get_default.

language

the new language tag.

fn set_matrix<'a, P: Into<Option<&'a Matrix>>>(&self, matrix: P)

Sets the transformation matrix that will be applied when rendering with this context. Note that reported metrics are in the user space coordinates before the application of the matrix, not device-space coordinates after the application of the matrix. So, they don't scale with the matrix, though they may change slightly for different matrices, depending on how the text is fit to the pixel grid.

matrix

a Matrix, or None to unset any existing matrix. (No matrix set is the same as setting the identity matrix.)

Loading content...