Trait pango::FontExt

pub trait FontExt: 'static {
    fn describe(&self) -> Option<FontDescription>;

    fn describe_with_absolute_size(&self) -> Option<FontDescription>;

    fn find_shaper(&self, language: &Language, ch: u32) -> Option<EngineShape>;

    fn coverage(&self, language: &Language) -> Option<Coverage>;

    fn face(&self) -> Option<FontFace>;

    fn font_map(&self) -> Option<FontMap>;

    fn glyph_extents(&self, glyph: Glyph) -> (Rectangle, Rectangle);

    fn metrics(&self, language: Option<&Language>) -> Option<FontMetrics>;

    fn has_char(&self, wc: char) -> bool;
}

Trait containing all Font methods.

Implementors

Font

Required methods

fn describe(&self) -> Option<FontDescription>

Returns a description of the font, with font size set in points. Use FontExt::describe_with_absolute_size if you want the font size in device units.

Returns

a newly-allocated FontDescription object.

fn describe_with_absolute_size(&self) -> Option<FontDescription>

Returns a description of the font, with absolute font size set (in device units). Use FontExt::describe if you want the font size in points.

Returns

a newly-allocated FontDescription object.

fn find_shaper(&self, language: &Language, ch: u32) -> Option<EngineShape>

Finds the best matching shaper for a font for a particular language tag and character point.

Deprecated

Shape engines are no longer used

language

the language tag

ch

a Unicode character.

Returns

the best matching shaper.

fn coverage(&self, language: &Language) -> Option<Coverage>

fn face(&self) -> Option<FontFace>

This is supported on crate feature v1_46 only.

fn font_map(&self) -> Option<FontMap>

fn glyph_extents(&self, glyph: Glyph) -> (Rectangle, Rectangle)

fn metrics(&self, language: Option<&Language>) -> Option<FontMetrics>

fn has_char(&self, wc: char) -> bool

This is supported on crate feature v1_44 only.

Returns whether the font provides a glyph for this character.

Returns true if self can render wc

Feature: v1_44

wc

a Unicode character

Loading content...