Crate pango

Bindings and wrappers for PANGO

Modules

ffi

widgets

Structs

FontDescription

FontMask

GlyphString
The PangoGlyphString structure is used to store strings of glyphs with geometry and visual attribute information. The storage for the glyph information is owned by the structure which simplifies memory management.

Item
The PangoItem structure stores information about a segment of text.

Enums

BidiType

Direction

Gravity

GravityHint

Script

Stretch

Style

Variant

Weight

Traits

Matrix
A structure specifying a transformation between user-space coordinates and device coordinates. The transformation is given by:

Rectangle
The PangoRectangle structure represents a rectangle. It is frequently used to represent the logical or ink extents of a single glyph or section of text. (See, for instance, pango_font_get_glyph_extents()).