Crate pango

Re-exports

pub use ffi;

pub use glib;

pub use crate::analysis::Analysis;

pub use crate::attr_class::AttrClass;

pub use crate::language::Language;

pub use crate::layout::HitPosition;

pub use crate::rectangle::Rectangle;

pub use glyph::GlyphGeometry;

pub use glyph::GlyphInfo;

Modules

analysis

attr_class

attr_iterator

attr_list

attribute

functions

glyph

item

language

layout

prelude
Traits and essential types inteded for blanket imports.

rectangle

Structs

AttrIterator
The crate::AttrIterator structure is used to represent an iterator through a crate::AttrList. A new iterator is created with [crate::AttrList::get_iterator()]. Once the iterator is created, it can be advanced through the style changes in the text using Self::next(). At each style change, the range of the current style segment and the attributes currently in effect can be queried.

AttrList
The crate::AttrList structure represents a list of attributes that apply to a section of text. The attributes are, in general, allowed to overlap in an arbitrary fashion, however, if the attributes are manipulated only through Self::change(), the overlap between properties will meet stricter criteria.

Attribute
The crate::Attribute structure represents the common portions of all attributes. Particular types of attributes include this structure as their initial portion. The common portion of the attribute holds the range to which the value in the type-specific part of the attribute applies and should be initialized using Self::init(). By default an attribute will have an all-inclusive range of [0,G_MAXUINT].

Color
The crate::Color structure is used to represent a color in an uncalibrated RGB color-space.

Context
The crate::Context structure stores global information used to control the itemization process.

Coverage
The crate::Coverage structure represents a map from Unicode characters to crate::CoverageLevel. It is an opaque structure with no public fields.

Font
The crate::Font structure is used to represent a font in a rendering-system-independent matter. To create an implementation of a crate::Font, the rendering-system specific code should allocate a larger structure that contains a nested crate::Font, fill in the <structfield>klass</structfield> member of the nested crate::Font with a pointer to a appropriate [crate::FontClass], then call pango_font_init() on the structure.

FontDescription
The crate::FontDescription structure represents the description of an ideal font. These structures are used both to list what fonts are available on the system and also for specifying the characteristics of a font to load.

FontFace
The crate::FontFace structure is used to represent a group of fonts with the same family, slant, weight, width, but varying sizes.

FontFamily
The crate::FontFamily structure is used to represent a family of related font faces. The faces in a family share a common design, but differ in slant, weight, width and other aspects.

FontMap
The crate::FontMap represents the set of fonts available for a particular rendering system. This is a virtual object with implementations being specific to particular rendering systems. To create an implementation of a crate::FontMap, the rendering-system specific code should allocate a larger structure that contains a nested crate::FontMap, fill in the <structfield>klass</structfield> member of the nested crate::FontMap with a pointer to a appropriate [crate::FontMapClass], then call pango_font_map_init() on the structure.

FontMask
The bits in a crate::FontMask correspond to fields in a crate::FontDescription that have been set.

FontMetrics
A crate::FontMetrics structure holds the overall metric information for a font (possibly restricted to a script). The fields of this structure are private to implementations of a font backend. See the documentation of the corresponding getters for documentation of their meaning.

Fontset
A crate::Fontset represents a set of crate::Font to use when rendering text. It is the result of resolving a crate::FontDescription against a particular crate::Context. It has operations for finding the component font for a particular Unicode character, and for finding a composite set of metrics for the entire fontset.

FontsetSimple
crate::FontsetSimple is a implementation of the abstract crate::Fontset base class in terms of an array of fonts, which the creator provides when constructing the crate::FontsetSimple.

GlyphItem
A crate::GlyphItem is a pair of a crate::Item and the glyphs resulting from shaping the text corresponding to an item. As an example of the usage of crate::GlyphItem, the results of shaping text with crate::Layout is a list of crate::LayoutLine, each of which contains a list of crate::GlyphItem.

GlyphItemIter
A crate::GlyphItemIter is an iterator over the clusters in a crate::GlyphItem. The <firstterm>forward direction</firstterm> of the iterator is the logical direction of text. That is, with increasing start_index and start_char values. If glyph_item is right-to-left (that is, if <literal>``glyph_item->item->analysis.level</literal> is odd), then start_glyph decreases as the iterator moves forward. Moreover, in right-to-left cases, start_glyph is greater than end_glyph.

GlyphString
The crate::GlyphString structure is used to store strings of glyphs with geometry and visual attribute information. The storage for the glyph information is owned by the structure which simplifies memory management.

Item
The crate::Item structure stores information about a segment of text.

Layout
The crate::Layout structure represents an entire paragraph of text. It is initialized with a crate::Context, UTF-8 string and set of attributes for that string. Once that is done, the set of formatted lines can be extracted from the object, the layout can be rendered, and conversion between logical character positions within the layout’s text, and the physical position of the resulting glyphs can be made.

LayoutIter
A crate::LayoutIter structure can be used to iterate over the visual extents of a crate::Layout.

LayoutLine
The crate::LayoutLine structure represents one of the lines resulting from laying out a paragraph via crate::Layout. crate::LayoutLine structures are obtained by calling [crate::Layout::get_line()] and are only valid until the text, attributes, or settings of the parent crate::Layout are modified.

Matrix
A structure specifying a transformation between user-space coordinates and device coordinates. The transformation is given by

Renderer
crate::Renderer is a base class for objects that are used to render Pango objects such as crate::GlyphString and crate::Layout.

ShapeFlags
Flags influencing the shaping process. These can be passed to pango_shape_with_flags().

ShowFlags
These flags affect how Pango treats characters that are normally not visible in the output.

TabArray
A crate::TabArray struct contains an array of tab stops. Each tab stop has an alignment and a position.

Enums

Alignment
A crate::Alignment describes how to align the lines of a crate::Layout within the available space. If the crate::Layout is set to justify using crate::Layout::set_justify(), this only has effect for partial lines.

AttrType
The crate::AttrType distinguishes between different types of attributes. Along with the predefined values, it is possible to allocate additional values for custom attributes using [Self::register()]. The predefined values are given below. The type of structure used to store the attribute is listed in parentheses after the description.

BidiType
The crate::BidiType type represents the bidirectional character type of a Unicode character as specified by the Unicode bidirectional algorithm</ulink>.

CoverageLevel
Used to indicate how well a font can represent a particular Unicode character point for a particular script.

Direction
The crate::Direction type represents a direction in the Unicode bidirectional algorithm; not every value in this enumeration makes sense for every usage of crate::Direction; for example, the return value of pango_unichar_direction() and pango_find_base_dir() cannot be Self::WeakLtr or Self::WeakRtl, since every character is either neutral or has a strong direction; on the other hand Self::Neutral doesn’t make sense to pass to pango_itemize_with_base_dir().

EllipsizeMode
The crate::EllipsizeMode type describes what sort of (if any) ellipsization should be applied to a line of text. In the ellipsization process characters are removed from the text in order to make it fit to a given width and replaced with an ellipsis.

Gravity
The crate::Gravity type represents the orientation of glyphs in a segment of text. This is useful when rendering vertical text layouts. In those situations, the layout is rotated using a non-identity PangoMatrix, and then glyph orientation is controlled using crate::Gravity. Not every value in this enumeration makes sense for every usage of crate::Gravity; for example, Self::Auto only can be passed to crate::Context::set_base_gravity() and can only be returned by [crate::Context::get_base_gravity()].

GravityHint
The crate::GravityHint defines how horizontal scripts should behave in a vertical context. That is, English excerpt in a vertical paragraph for example.

Overline	v1_46
The crate::Overline enumeration is used to specify whether text should be overlined, and if so, the type of line.

RenderPart
crate::RenderPart defines different items to render for such purposes as setting colors.

Script
The crate::Script enumeration identifies different writing systems. The values correspond to the names as defined in the Unicode standard. See Unicode Standard Annex 24: Script names</ulink>.

Stretch
An enumeration specifying the width of the font relative to other designs within a family.

Style
An enumeration specifying the various slant styles possible for a font.

TabAlign
A crate::TabAlign specifies where a tab stop appears relative to the text.

Underline
The crate::Underline enumeration is used to specify whether text should be underlined, and if so, the type of underlining.

Variant
An enumeration specifying capitalization variant of the font.

Weight
An enumeration specifying the weight (boldness) of a font. This is a numerical value ranging from 100 to 1000, but there are some predefined values:

WrapMode
A crate::WrapMode describes how to wrap the lines of a crate::Layout to the desired width.

Constants

NONE_FONT

NONE_FONTSET

NONE_FONT_FACE

NONE_FONT_FAMILY

NONE_FONT_MAP

NONE_RENDERER

SCALE

SCALE_LARGE
The scale factor for one magnification step (1.2).

SCALE_MEDIUM
The scale factor for normal size (1.0).

SCALE_SMALL
The scale factor for one shrinking step (1 / 1.2).

SCALE_XX_LARGE
The scale factor for three magnification steps (1.2 * 1.2 * 1.2).

SCALE_XX_SMALL
The scale factor for three shrinking steps (1 / (1.2 * 1.2 * 1.2)).

SCALE_X_LARGE
The scale factor for two magnification steps (1.2 * 1.2).

SCALE_X_SMALL
The scale factor for two shrinking steps (1 / (1.2 * 1.2)).

Functions

extents_to_pixels

find_base_dir

find_paragraph_boundary

is_zero_width

itemize

itemize_with_base_dir

parse_markup

parse_stretch

parse_style

parse_variant

parse_weight

quantize_line_geometry

reorder_items

shape

shape_full

shape_with_flags	v1_44

unichar_direction

units_from_double

units_to_double

version

version_check

version_string

Type Definitions

Glyph

GlyphUnit

LayoutRun