Enum pango::Script

#[non_exhaustive]
pub enum Script {

    InvalidCode,
    Common,
    Inherited,
    Arabic,
    Armenian,
    Bengali,
    Bopomofo,
    Cherokee,
    Coptic,
    Cyrillic,
    Deseret,
    Devanagari,
    Ethiopic,
    Georgian,
    Gothic,
    Greek,
    Gujarati,
    Gurmukhi,
    Han,
    Hangul,
    Hebrew,
    Hiragana,
    Kannada,
    Katakana,
    Khmer,
    Lao,
    Latin,
    Malayalam,
    Mongolian,
    Myanmar,
    Ogham,
    OldItalic,
    Oriya,
    Runic,
    Sinhala,
    Syriac,
    Tamil,
    Telugu,
    Thaana,
    Thai,
    Tibetan,
    CanadianAboriginal,
    Yi,
    Tagalog,
    Hanunoo,
    Buhid,
    Tagbanwa,
    Braille,
    Cypriot,
    Limbu,
    Osmanya,
    Shavian,
    LinearB,
    TaiLe,
    Ugaritic,
    NewTaiLue,
    Buginese,
    Glagolitic,
    Tifinagh,
    SylotiNagri,
    OldPersian,
    Kharoshthi,
    Unknown,
    Balinese,
    Cuneiform,
    Phoenician,
    PhagsPa,
    Nko,
    KayahLi,
    Lepcha,
    Rejang,
    Sundanese,
    Saurashtra,
    Cham,
    OlChiki,
    Vai,
    Carian,
    Lycian,
    Lydian,
    Batak,
    Brahmi,
    Mandaic,
    Chakma,
    MeroiticCursive,
    MeroiticHieroglyphs,
    Miao,
    Sharada,
    SoraSompeng,
    Takri,
    BassaVah,
    CaucasianAlbanian,
    Duployan,
    Elbasan,
    Grantha,
    Khojki,
    Khudawadi,
    LinearA,
    Mahajani,
    Manichaean,
    MendeKikakui,
    Modi,
    Mro,
    Nabataean,
    OldNorthArabian,
    OldPermic,
    PahawhHmong,
    Palmyrene,
    PauCinHau,
    PsalterPahlavi,
    Siddham,
    Tirhuta,
    WarangCiti,
    Ahom,
    AnatolianHieroglyphs,
    Hatran,
    Multani,
    OldHungarian,
    Signwriting,

}

The Script enumeration identifies different writing systems.

The values correspond to the names as defined in the Unicode standard. See Unicode Standard Annex 24: Script names

Note that this enumeration is deprecated and will not be updated to include values in newer versions of the Unicode standard. Applications should use the GLib::UnicodeScript enumeration instead, whose values are interchangeable with Script.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

InvalidCode

a value never returned from pango_script_for_unichar()

Common

a character used by multiple different scripts

Inherited

a mark glyph that takes its script from the base glyph to which it is attached

Arabic

Arabic

Armenian

Armenian

Bengali

Bengali

Bopomofo

Bopomofo

Cherokee

Cherokee

Coptic

Coptic

Cyrillic

Cyrillic

Deseret

Deseret

Devanagari

Devanagari

Ethiopic

Ethiopic

Georgian

Georgian

Gothic

Gothic

Greek

Greek

Gujarati

Gujarati

Gurmukhi

Gurmukhi

Han

Han

Hangul

Hangul

Hebrew

Hebrew

Hiragana

Hiragana

Kannada

Kannada

Katakana

Katakana

Khmer

Khmer

Lao

Lao

Latin

Latin

Malayalam

Malayalam

Mongolian

Mongolian

Myanmar

Myanmar

Ogham

Ogham

OldItalic

Old Italic

Oriya

Oriya

Runic

Runic

Sinhala

Sinhala

Syriac

Syriac

Tamil

Tamil

Telugu

Telugu

Thaana

Thaana

Thai

Thai

Tibetan

Tibetan

CanadianAboriginal

Canadian Aboriginal

Yi

Yi

Tagalog

Tagalog

Hanunoo

Hanunoo

Buhid

Buhid

Tagbanwa

Tagbanwa

Braille

Braille

Cypriot

Cypriot

Limbu

Limbu

Osmanya

Osmanya

Shavian

Shavian

LinearB

Linear B

TaiLe

Tai Le

Ugaritic

Ugaritic

NewTaiLue

New Tai Lue. Since 1.10

Buginese

Buginese. Since 1.10

Glagolitic

Glagolitic. Since 1.10

Tifinagh

Tifinagh. Since 1.10

SylotiNagri

Syloti Nagri. Since 1.10

OldPersian

Old Persian. Since 1.10

Kharoshthi

Kharoshthi. Since 1.10

Unknown

an unassigned code point. Since 1.14

Balinese

Balinese. Since 1.14

Cuneiform

Cuneiform. Since 1.14

Phoenician

Phoenician. Since 1.14

PhagsPa

Phags-pa. Since 1.14

Nko

N’Ko. Since 1.14

KayahLi

Kayah Li. Since 1.20.1

Lepcha

Lepcha. Since 1.20.1

Rejang

Rejang. Since 1.20.1

Sundanese

Sundanese. Since 1.20.1

Saurashtra

Saurashtra. Since 1.20.1

Cham

Cham. Since 1.20.1

OlChiki

Ol Chiki. Since 1.20.1

Vai

Vai. Since 1.20.1

Carian

Carian. Since 1.20.1

Lycian

Lycian. Since 1.20.1

Lydian

Lydian. Since 1.20.1

Batak

Batak. Since 1.32

Brahmi

Brahmi. Since 1.32

Mandaic

Mandaic. Since 1.32

Chakma

Chakma. Since: 1.32

MeroiticCursive

Meroitic Cursive. Since: 1.32

MeroiticHieroglyphs

Meroitic Hieroglyphs. Since: 1.32

Miao

Miao. Since: 1.32

Sharada

Sharada. Since: 1.32

SoraSompeng

Sora Sompeng. Since: 1.32

Takri

Takri. Since: 1.32

BassaVah

Bassa. Since: 1.40

CaucasianAlbanian

Caucasian Albanian. Since: 1.40

Duployan

Duployan. Since: 1.40

Elbasan

Elbasan. Since: 1.40

Grantha

Grantha. Since: 1.40

Khojki

Kjohki. Since: 1.40

Khudawadi

Khudawadi, Sindhi. Since: 1.40

LinearA

Linear A. Since: 1.40

Mahajani

Mahajani. Since: 1.40

Manichaean

Manichaean. Since: 1.40

MendeKikakui

Mende Kikakui. Since: 1.40

Modi

Modi. Since: 1.40

Mro

Mro. Since: 1.40

Nabataean

Nabataean. Since: 1.40

OldNorthArabian

Old North Arabian. Since: 1.40

OldPermic

Old Permic. Since: 1.40

PahawhHmong

Pahawh Hmong. Since: 1.40

Palmyrene

Palmyrene. Since: 1.40

PauCinHau

Pau Cin Hau. Since: 1.40

PsalterPahlavi

Psalter Pahlavi. Since: 1.40

Siddham

Siddham. Since: 1.40

Tirhuta

Tirhuta. Since: 1.40

WarangCiti

Warang Citi. Since: 1.40

Ahom

Ahom. Since: 1.40

AnatolianHieroglyphs

Anatolian Hieroglyphs. Since: 1.40

Hatran

Hatran. Since: 1.40

Multani

Multani. Since: 1.40

OldHungarian

Old Hungarian. Since: 1.40

Signwriting

Signwriting. Since: 1.40

Implementations

impl Script

pub fn for_unichar(ch: char) -> Script

pub fn sample_language(self) -> Option<Language>