Function pango::itemize_with_base_dir

pub fn itemize_with_base_dir(
    context: &Context,
    base_dir: Direction,
    text: &str,
    start_index: i32,
    length: i32,
    attrs: &AttrList,
    cached_iter: Option<&AttrIterator<'_>>,
) -> Vec<Item>

Like pango_itemize(), but with an explicitly specified base direction.

The base direction is used when computing bidirectional levels. itemize() gets the base direction from the Context (see Context::set_base_dir()).

context

a structure holding information that affects the itemization process.

base_dir

base direction to use for bidirectional processing

text

the text to itemize.

start_index

first byte in @text to process

length

the number of bytes (not characters) to process after @start_index. This must be >= 0.

attrs

the set of attributes that apply to @text.

cached_iter

Cached attribute iterator

Returns

a GList of Item structures. The items should be freed using Pango::Item::free() probably in combination with GLib::List::free_full().