Function pango::itemize_with_base_dir

pub fn itemize_with_base_dir(
    context: &Context, 
    base_dir: Direction, 
    text: &str, 
    start_index: i32, 
    length: i32, 
    attrs: &AttrList, 
    cached_iter: Option<&AttrIterator>
) -> Vec<Item>

Like itemize(), but the base direction to use when computing bidirectional levels (see pango_context_set_base_dir ()), is specified explicitly rather than gotten from the Context.

context

a structure holding information that affects the itemization process.

base_dir

base direction to use for bidirectional processing

text

the text to itemize.

start_index

first byte in text to process

length

the number of bytes (not characters) to process after start_index. This must be >= 0.

attrs

the set of attributes that apply to text.

cached_iter

Cached attribute iterator, or None

Returns

a GList of Item structures. The items should be freed using pango_item_free() probably in combination with g_list_foreach(), and the list itself using g_list_free().