pango

Function markup_parser_new

pub fn markup_parser_new(accel_marker: char) -> MarkupParseContext

Incrementally parses marked-up text to create a plain-text string and an attribute list.

See the Pango Markup docs for details about the supported markup.

If @accel_marker is nonzero, the given character will mark the character following it as an accelerator. For example, @accel_marker might be an ampersand or underscore. All characters marked as an accelerator will receive a Underline::Low attribute, and the first character so marked will be returned in @accel_char, when calling markup_parser_finish(). Two @accel_marker characters following each other produce a single literal @accel_marker character.

To feed markup to the parser, use glib::MarkupParseContext::parse() on the returned glib::MarkupParseContext. When done with feeding markup to the parser, use markup_parser_finish() to get the data out of it, and then use GLib::MarkupParseContext::free() to free it.

This function is designed for applications that read Pango markup from streams. To simply parse a string containing Pango markup, the parse_markup() API is recommended instead.

accel_marker

character that precedes an accelerator, or 0 for none

Returns

a GMarkupParseContext that should be destroyed with GLib::MarkupParseContext::free().