Module pango::functions

Functions

find_base_dir

Searches a string the first character that has a strong direction, according to the Unicode bidirectional algorithm.

find_paragraph_boundary

Locates a paragraph boundary in @text.

is_zero_width

Checks if a character that should not be normally rendered.

itemize

Breaks a piece of text into segments with consistent directional level and font.

itemize_with_base_dir

Like pango_itemize(), but with an explicitly specified base direction.

markup_parser_finish

Finishes parsing markup.

markup_parser_new

Incrementally parses marked-up text to create a plain-text string and an attribute list.

parse_markup

Parses marked-up text to create a plain-text string and an attribute list.

parse_stretch

Parses a font stretch.

parse_style

Parses a font style.

parse_variant

Parses a font variant.

parse_weight

Parses a font weight.

quantize_line_geometry

Quantizes the thickness and position of a line to whole device pixels.

shape

Convert the characters in @text into glyphs.

unichar_direction

Determines the inherent direction of a character.

units_from_double

Converts a floating-point number to Pango units.

units_to_double

Converts a number in Pango units to floating-point.

version

Returns the encoded version of Pango available at run-time.

version_check

Checks that the Pango library in use is compatible with the given version.

version_string

Returns the version of Pango available at run-time.