Trait pango::prelude::FontsetExt

pub trait FontsetExt: 'static {
    // Required methods
    fn foreach<P: FnMut(&Fontset, &Font) -> bool>(&self, func: P);
    fn font(&self, wc: u32) -> Option<Font>;
    fn metrics(&self) -> Option<FontMetrics>;
}

Trait containing all Fontset methods.

Implementors

FontsetSimple, Fontset

Required Methods

fn foreach<P: FnMut(&Fontset, &Font) -> bool>(&self, func: P)

Iterates through all the fonts in a fontset, calling @func for each one.

If @func returns true, that stops the iteration.

func

Callback function

fn font(&self, wc: u32) -> Option<Font>

Returns the font in the fontset that contains the best glyph for a Unicode character.

wc

a Unicode character

Returns

a Font

fn metrics(&self) -> Option<FontMetrics>

Get overall metric information for the fonts in the fontset.

Returns

a FontMetrics object