pango

Struct AttrList

pub struct AttrList { /* private fields */ }

A AttrList represents a list of attributes that apply to a section of text.

The attributes in a AttrList are, in general, allowed to overlap in an arbitrary fashion. However, if the attributes are manipulated only through change(), the overlap between properties will meet stricter criteria.

Since the AttrList structure is stored as a linear list, it is not suitable for storing attributes for large amounts of text. In general, you should not use a single AttrList for more than one paragraph of text.

GLib type: Shared boxed type with reference counted clone semantics.

Implementations

impl AttrList

pub fn as_ptr(&self) -> *mut PangoAttrList

Return the inner pointer to the underlying C value.

pub unsafe fn from_glib_ptr_borrow(ptr: &*mut PangoAttrList) -> &Self

Borrows the underlying C value.

impl AttrList

pub fn new() -> AttrList

Create a new empty attribute list with a reference count of one.

Returns

the newly allocated AttrList, which should be freed with Pango::AttrList::unref()

pub fn copy(&self) -> Option<AttrList>

pub fn filter<P: FnMut(&Attribute) -> bool>(&self, func: P) -> Option<AttrList>

Given a AttrList and callback function, removes any elements of @self for which @func returns true and inserts them into a new list.

func

callback function; returns true if an attribute should be filtered out

Returns

the new AttrList or None if no attributes of the given types were found

pub fn attributes(&self) -> Vec<Attribute>

Available on crate feature v1_44 only.

Gets a list of all attributes in @self.

Returns

a list of all attributes in @self. To free this value, call Pango::Attribute::destroy() on each value and g_slist_free() on the list.

pub fn splice(&self, other: &AttrList, pos: i32, len: i32)

This function opens up a hole in @self, fills it in with attributes from the left, and then merges @other on top of the hole.

This operation is equivalent to stretching every attribute that applies at position @pos in @self by an amount @len, and then calling change() with a copy of each attribute in @other in sequence (offset in position by @pos, and limited in length to @len).

This operation proves useful for, for instance, inserting a pre-edit string in the middle of an edit buffer.

For backwards compatibility, the function behaves differently when @len is 0. In this case, the attributes from @other are not imited to @len, and are just overlayed on top of @self.

This mode is useful for merging two lists of attributes together.

other

another AttrList

pos

the position in @self at which to insert @other

len

the length of the spliced segment. (Note that this must be specified since the attributes in @other may only be present at some subsection of this range)

pub fn to_str(&self) -> GString

Available on crate feature v1_50 only.

Serializes a AttrList to a string.

In the resulting string, serialized attributes are separated by newlines or commas. Individual attributes are serialized to a string of the form

[START END] TYPE VALUE

Where START and END are the indices (with -1 being accepted in place of MAXUINT), TYPE is the nickname of the attribute value type, e.g. weight or stretch, and the value is serialized according to its type:

Optionally, START and END can be omitted to indicate unlimited extent.

enum values as nick or numeric value

boolean values as true or false

integers and floats as numbers

strings as string, optionally quoted

font features as quoted string

PangoLanguage as string

PangoFontDescription as serialized by FontDescription::to_str(), quoted

PangoColor as serialized by Color::to_str()

Examples:

0 10 foreground red, 5 15 weight bold, 0 200 font-desc "Sans 10"

0 -1 weight 700
0 100 family Times

weight bold

To parse the returned value, use from_string().

Note that shape attributes can not be serialized.

Returns

a newly allocated string

pub fn update(&self, pos: i32, remove: i32, add: i32)

Available on crate feature v1_44 only.

Update indices of attributes in @self for a change in the text they refer to.

The change that this function applies is removing @remove bytes at position @pos and inserting @add bytes instead.

Attributes that fall entirely in the (@pos, @pos + @remove) range are removed.

Attributes that start or end inside the (@pos, @pos + @remove) range are shortened to reflect the removal.

Attributes start and end positions are updated if they are behind @pos + @remove.

pos

the position of the change

remove

the number of removed bytes

add

the number of added bytes

pub fn from_string(text: &str) -> Result<AttrList, BoolError>

Available on crate feature v1_50 only.

Deserializes a AttrList from a string.

This is the counterpart to to_str(). See that functions for details about the format.

text

a string

Returns

a new AttrList

impl AttrList

pub fn change(&self, attr: impl Into<Attribute>)

Insert the given attribute into the AttrList.

It will replace any attributes of the same type on that segment and be merged with any adjoining attributes that are identical.

This function is slower than insert() for creating an attribute list in order (potentially much slower for large lists). However, insert() is not suitable for continually changing a set of attributes since it never removes or combines existing attributes.

attr

the attribute to insert

pub fn insert(&self, attr: impl Into<Attribute>)

Insert the given attribute into the AttrList.

It will be inserted after all other attributes with a matching @start_index.

attr

the attribute to insert

pub fn insert_before(&self, attr: impl Into<Attribute>)

Insert the given attribute into the AttrList.

It will be inserted before all other attributes with a matching @start_index.

attr

the attribute to insert

pub fn iterator(&self) -> AttrIterator<'_>

Create a iterator initialized to the beginning of the list.

@self must not be modified until this iterator is freed.

Returns

the newly allocated AttrIterator, which should be freed with Pango::AttrIterator::destroy()