Struct pango::Layout

pub struct Layout(_, _);

The Layout structure represents an entire paragraph of text. It is initialized with a Context, UTF-8 string and set of attributes for that string. Once that is done, the set of formatted lines can be extracted from the object, the layout can be rendered, and conversion between logical character positions within the layout's text, and the physical position of the resulting glyphs can be made.

There are also a number of parameters to adjust the formatting of a Layout, which are illustrated in . It is possible, as well, to ignore the 2-D setup, and simply treat the results of a Layout as a list of lines.
``Adjustable parameters for a PangoLayout`` `` ``
The Layout structure is opaque, and has no user-visible fields.

Implements

LayoutExt

Methods

impl Layout

pub fn new(context: &Context) -> Layout

Create a new Layout object with attributes initialized to default values for a particular Context.

context

a Context

Returns

the newly allocated Layout, with a reference count of one, which should be freed with gobject::Object::unref.

pub fn context_changed(&self)

pub fn copy(&self) -> Option<Layout>

pub fn get_alignment(&self) -> Alignment

pub fn get_attributes(&self) -> Option<AttrList>

pub fn get_auto_dir(&self) -> bool

pub fn get_baseline(&self) -> i32

pub fn get_character_count(&self) -> i32

pub fn get_context(&self) -> Option<Context>

pub fn get_cursor_pos(&self, index_: i32) -> (Rectangle, Rectangle)

pub fn get_ellipsize(&self) -> EllipsizeMode

pub fn get_extents(&self) -> (Rectangle, Rectangle)

pub fn get_font_description(&self) -> Option<FontDescription>

pub fn get_height(&self) -> i32

pub fn get_indent(&self) -> i32

pub fn get_iter(&self) -> Option<LayoutIter>

pub fn get_justify(&self) -> bool

pub fn get_line(&self, line: i32) -> Option<LayoutLine>

pub fn get_line_count(&self) -> i32

pub fn get_line_readonly(&self, line: i32) -> Option<LayoutLine>

pub fn get_lines(&self) -> Vec<LayoutLine>

pub fn get_lines_readonly(&self) -> Vec<LayoutLine>

pub fn get_pixel_extents(&self) -> (Rectangle, Rectangle)

pub fn get_pixel_size(&self) -> (i32, i32)

pub fn get_serial(&self) -> u32

pub fn get_single_paragraph_mode(&self) -> bool

pub fn get_size(&self) -> (i32, i32)

pub fn get_spacing(&self) -> i32

pub fn get_tabs(&self) -> Option<TabArray>

pub fn get_text(&self) -> Option<GString>

pub fn get_unknown_glyphs_count(&self) -> i32

pub fn get_width(&self) -> i32

pub fn get_wrap(&self) -> WrapMode

pub fn index_to_line_x(&self, index_: i32, trailing: bool) -> (i32, i32)

pub fn index_to_pos(&self, index_: i32) -> Rectangle

pub fn is_ellipsized(&self) -> bool

pub fn is_wrapped(&self) -> bool

pub fn move_cursor_visually(
&self,
strong: bool,
old_index: i32,
old_trailing: i32,
direction: i32
) -> (i32, i32)

pub fn set_alignment(&self, alignment: Alignment)

pub fn set_attributes(&self, attrs: Option<&AttrList>)

pub fn set_auto_dir(&self, auto_dir: bool)

pub fn set_ellipsize(&self, ellipsize: EllipsizeMode)

pub fn set_font_description(&self, desc: Option<&FontDescription>)

pub fn set_height(&self, height: i32)

pub fn set_indent(&self, indent: i32)

pub fn set_justify(&self, justify: bool)

pub fn set_markup(&self, markup: &str)

pub fn set_markup_with_accel(&self, markup: &str, accel_marker: char) -> char

pub fn set_single_paragraph_mode(&self, setting: bool)

pub fn set_spacing(&self, spacing: i32)

pub fn set_tabs(&self, tabs: Option<&TabArray>)

pub fn set_text(&self, text: &str)

pub fn set_width(&self, width: i32)

pub fn set_wrap(&self, wrap: WrapMode)

pub fn xy_to_index(&self, x: i32, y: i32) -> (bool, i32, i32)