Struct pango::Layout

pub struct Layout(_);

The crate::Layout structure represents an entire paragraph of text. It is initialized with a crate::Context, UTF-8 string and set of attributes for that string. Once that is done, the set of formatted lines can be extracted from the object, the layout can be rendered, and conversion between logical character positions within the layout’s text, and the physical position of the resulting glyphs can be made.

There are also a number of parameters to adjust the formatting of a crate::Layout, which are illustrated in . It is possible, as well, to ignore the 2-D setup, and simply treat the results of a crate::Layout as a list of lines.
``Adjustable parameters (on the left) and font metrics (on the right) for a PangoLayout`` `` ``
The crate::Layout structure is opaque, and has no user-visible fields.

Implementations

impl Layout

pub fn new(context: &Context) -> Layout

Create a new crate::Layout object with attributes initialized to default values for a particular crate::Context.

context

a crate::Context

Returns

the newly allocated crate::Layout, with a reference count of one, which should be freed with [crate::glib::object::Object::unref()].

pub fn context_changed(&self)

Forces recomputation of any state in the crate::Layout that might depend on the layout’s context. This function should be called if you make changes to the context subsequent to creating the layout.

pub fn copy(&self) -> Option<Layout>

Does a deep copy-by-value of the self layout. The attribute list, tab array, and text from the original layout are all copied by value.

Returns

the newly allocated crate::Layout, with a reference count of one, which should be freed with [crate::glib::object::Object::unref()].

pub fn alignment(&self) -> Alignment

Gets the alignment for the layout: how partial lines are positioned within the horizontal space available.

Returns

the alignment.

pub fn attributes(&self) -> Option<AttrList>

Gets the attribute list for the layout, if any.

Returns

a crate::AttrList or None if none was set.

pub fn is_auto_dir(&self) -> bool

Gets whether to calculate the bidirectional base direction for the layout according to the contents of the layout. See Self::set_auto_dir().

Returns

true if the bidirectional base direction is computed from the layout’s contents, false otherwise.

pub fn baseline(&self) -> i32

Gets the Y position of baseline of the first line in self.

Returns

baseline of first line, from top of self.

pub fn character_count(&self) -> i32

Returns the number of Unicode characters in the the text of self.

Returns

the number of Unicode characters in the text of self

pub fn context(&self) -> Option<Context>

Retrieves the crate::Context used for this layout.

Returns

the crate::Context for the layout. This does not have an additional refcount added, so if you want to keep a copy of this around, you must reference it yourself.

pub fn cursor_pos(&self, index_: i32) -> (Rectangle, Rectangle)

Given an index within a layout, determines the positions that of the strong and weak cursors if the insertion point is at that index. The position of each cursor is stored as a zero-width rectangle. The strong cursor location is the location where characters of the directionality equal to the base direction of the layout are inserted. The weak cursor location is the location where characters of the directionality opposite to the base direction of the layout are inserted.

index_

the byte index of the cursor

strong_pos

location to store the strong cursor position (may be None)

weak_pos

location to store the weak cursor position (may be None)

pub fn direction(&self, index: i32) -> Direction

This is supported on crate feature v1_46 only.

Gets the text direction at the given character position in self.

index

the byte index of the char

Returns

the text direction at index

pub fn ellipsize(&self) -> EllipsizeMode

Gets the type of ellipsization being performed for self. See Self::set_ellipsize()

Returns

the current ellipsization mode for self.

Use Self::is_ellipsized() to query whether any paragraphs were actually ellipsized.

pub fn extents(&self) -> (Rectangle, Rectangle)

Computes the logical and ink extents of self. Logical extents are usually what you want for positioning things. Note that both extents may have non-zero x and y. You may want to use those to offset where you render the layout. Not doing that is a very typical bug that shows up as right-to-left layouts not being correctly positioned in a layout with a set width.

The extents are given in layout coordinates and in Pango units; layout coordinates begin at the top left corner of the layout.

ink_rect

rectangle used to store the extents of the layout as drawn or None to indicate that the result is not needed.

logical_rect

rectangle used to store the logical extents of the layout or None to indicate that the result is not needed.

pub fn font_description(&self) -> Option<FontDescription>

Gets the font description for the layout, if any.

Returns

a pointer to the layout’s font description, or None if the font description from the layout’s context is inherited. This value is owned by the layout and must not be modified or freed.

pub fn height(&self) -> i32

Gets the height of layout used for ellipsization. See Self::set_height() for details.

Returns

the height, in Pango units if positive, or number of lines if negative.

pub fn indent(&self) -> i32

Gets the paragraph indent width in Pango units. A negative value indicates a hanging indentation.

Returns

the indent in Pango units.

pub fn iter(&self) -> Option<LayoutIter>

Returns an iterator to iterate over the visual extents of the layout.

Returns

the new crate::LayoutIter that should be freed using [crate::LayoutIter::free()].

pub fn is_justify(&self) -> bool

Gets whether each complete line should be stretched to fill the entire width of the layout.

Returns

the justify.

pub fn line(&self, line: i32) -> Option<LayoutLine>

Retrieves a particular line from a crate::Layout.

Use the faster [Self::get_line_readonly()] if you do not plan to modify the contents of the line (glyphs, glyph widths, etc.).

line

the index of a line, which must be between 0 and <literal>pango_layout_get_line_count(layout) - 1</literal>, inclusive.

Returns

the requested crate::LayoutLine, or None if the index is out of range. This layout line can be ref’ed and retained, but will become invalid if changes are made to the crate::Layout.

pub fn line_count(&self) -> i32

Retrieves the count of lines for the self.

Returns

the line count.

pub fn line_readonly(&self, line: i32) -> Option<LayoutLine>

Retrieves a particular line from a crate::Layout.

This is a faster alternative to [Self::get_line()], but the user is not expected to modify the contents of the line (glyphs, glyph widths, etc.).

line

the index of a line, which must be between 0 and <literal>pango_layout_get_line_count(layout) - 1</literal>, inclusive.

Returns

the requested crate::LayoutLine, or None if the index is out of range. This layout line can be ref’ed and retained, but will become invalid if changes are made to the crate::Layout. No changes should be made to the line.

pub fn line_spacing(&self) -> f32

This is supported on crate feature v1_44 only.

Gets the value that has been set with Self::set_line_spacing().

pub fn lines(&self) -> Vec<LayoutLine>

Returns the lines of the self as a list.

Use the faster [Self::get_lines_readonly()] if you do not plan to modify the contents of the lines (glyphs, glyph widths, etc.).

Returns

a [crate::glib::SList] containing the lines in the layout. This points to internal data of the crate::Layout and must be used with care. It will become invalid on any change to the layout’s text or properties.

pub fn lines_readonly(&self) -> Vec<LayoutLine>

Returns the lines of the self as a list.

This is a faster alternative to [Self::get_lines()], but the user is not expected to modify the contents of the lines (glyphs, glyph widths, etc.).

Returns

a [crate::glib::SList] containing the lines in the layout. This points to internal data of the crate::Layout and must be used with care. It will become invalid on any change to the layout’s text or properties. No changes should be made to the lines.

pub fn pixel_extents(&self) -> (Rectangle, Rectangle)

Computes the logical and ink extents of self in device units. This function just calls [Self::get_extents()] followed by two pango_extents_to_pixels() calls, rounding ink_rect and logical_rect such that the rounded rectangles fully contain the unrounded one (that is, passes them as first argument to pango_extents_to_pixels()).

ink_rect

rectangle used to store the extents of the layout as drawn or None to indicate that the result is not needed.

logical_rect

rectangle used to store the logical extents of the layout or None to indicate that the result is not needed.

pub fn pixel_size(&self) -> (i32, i32)

Determines the logical width and height of a crate::Layout in device units. ([Self::get_size()] returns the width and height scaled by PANGO_SCALE.) This is simply a convenience function around [Self::get_pixel_extents()].

width

location to store the logical width, or None

height

location to store the logical height, or None

pub fn serial(&self) -> u32

Returns the current serial number of self. The serial number is initialized to an small number larger than zero when a new layout is created and is increased whenever the layout is changed using any of the setter functions, or the crate::Context it uses has changed. The serial may wrap, but will never have the value 0. Since it can wrap, never compare it with “less than”, always use “not equals”.

This can be used to automatically detect changes to a crate::Layout, and is useful for example to decide whether a layout needs redrawing. To force the serial to be increased, use Self::context_changed().

Returns

The current serial number of self.

pub fn is_single_paragraph_mode(&self) -> bool

Obtains the value set by Self::set_single_paragraph_mode().

Returns

true if the layout does not break paragraphs at paragraph separator characters, false otherwise.

pub fn size(&self) -> (i32, i32)

Determines the logical width and height of a crate::Layout in Pango units (device units scaled by PANGO_SCALE). This is simply a convenience function around [Self::get_extents()].

width

location to store the logical width, or None

height

location to store the logical height, or None

pub fn spacing(&self) -> i32

Gets the amount of spacing between the lines of the layout.

Returns

the spacing in Pango units.

pub fn tabs(&self) -> Option<TabArray>

Gets the current crate::TabArray used by this layout. If no crate::TabArray has been set, then the default tabs are in use and None is returned. Default tabs are every 8 spaces. The return value should be freed with [crate::TabArray::free()].

Returns

a copy of the tabs for this layout, or None.

pub fn text(&self) -> Option<GString>

Gets the text in the layout. The returned text should not be freed or modified.

Returns

the text in the self.

pub fn unknown_glyphs_count(&self) -> i32

Counts the number unknown glyphs in self. That is, zero if glyphs for all characters in the layout text were found, or more than zero otherwise.

This function can be used to determine if there are any fonts available to render all characters in a certain string, or when used in combination with crate::AttrType::Fallback, to check if a certain font supports all the characters in the string.

Returns

The number of unknown glyphs in self.

pub fn width(&self) -> i32

Gets the width to which the lines of the crate::Layout should wrap.

Returns

the width in Pango units, or -1 if no width set.

pub fn wrap(&self) -> WrapMode

Gets the wrap mode for the layout.

Use Self::is_wrapped() to query whether any paragraphs were actually wrapped.

Returns

active wrap mode.

pub fn index_to_line_x(&self, index_: i32, trailing: bool) -> (i32, i32)

Converts from byte index_ within the self to line and X position. (X position is measured from the left edge of the line)

index_

the byte index of a grapheme within the layout.

trailing

an integer indicating the edge of the grapheme to retrieve the position of. If > 0, the trailing edge of the grapheme, if 0, the leading of the grapheme.

line

location to store resulting line index. (which will between 0 and pango_layout_get_line_count(layout) - 1), or None

x_pos

location to store resulting position within line (PANGO_SCALE units per device unit), or None

pub fn index_to_pos(&self, index_: i32) -> Rectangle

Converts from an index within a crate::Layout to the onscreen position corresponding to the grapheme at that index, which is represented as rectangle. Note that <literal>pos->x</literal> is always the leading edge of the grapheme and <literal>pos->x + pos->width</literal> the trailing edge of the grapheme. If the directionality of the grapheme is right-to-left, then <literal>pos->width</literal> will be negative.

index_

byte index within self

pos

rectangle in which to store the position of the grapheme

pub fn is_ellipsized(&self) -> bool

Queries whether the layout had to ellipsize any paragraphs.

This returns true if the ellipsization mode for self is not crate::EllipsizeMode::None, a positive width is set on self, and there are paragraphs exceeding that width that have to be ellipsized.

Returns

true if any paragraphs had to be ellipsized, false otherwise.

pub fn is_wrapped(&self) -> bool

Queries whether the layout had to wrap any paragraphs.

This returns true if a positive width is set on self, ellipsization mode of self is set to crate::EllipsizeMode::None, and there are paragraphs exceeding the layout width that have to be wrapped.

Returns

true if any paragraphs had to be wrapped, false otherwise.

pub fn move_cursor_visually(
&self,
strong: bool,
old_index: i32,
old_trailing: i32,
direction: i32
) -> (i32, i32)

Computes a new cursor position from an old position and a count of positions to move visually. If direction is positive, then the new strong cursor position will be one position to the right of the old cursor position. If direction is negative, then the new strong cursor position will be one position to the left of the old cursor position.

In the presence of bidirectional text, the correspondence between logical and visual order will depend on the direction of the current run, and there may be jumps when the cursor is moved off of the end of a run.

Motion here is in cursor positions, not in characters, so a single call to Self::move_cursor_visually() may move the cursor over multiple characters when multiple characters combine to form a single grapheme.

strong

whether the moving cursor is the strong cursor or the weak cursor. The strong cursor is the cursor corresponding to text insertion in the base direction for the layout.

old_index

the byte index of the grapheme for the old index

old_trailing

if 0, the cursor was at the leading edge of the grapheme indicated by old_index, if > 0, the cursor was at the trailing edge.

direction

direction to move cursor. A negative value indicates motion to the left.

new_index

location to store the new cursor byte index. A value of -1 indicates that the cursor has been moved off the beginning of the layout. A value of G_MAXINT indicates that the cursor has been moved off the end of the layout.

new_trailing

number of characters to move forward from the location returned for new_index to get the position where the cursor should be displayed. This allows distinguishing the position at the beginning of one line from the position at the end of the preceding line. new_index is always on the line where the cursor should be displayed.

pub fn set_alignment(&self, alignment: Alignment)

Sets the alignment for the layout: how partial lines are positioned within the horizontal space available.

alignment

the alignment

pub fn set_attributes(&self, attrs: Option<&AttrList>)

Sets the text attributes for a layout object. References attrs, so the caller can unref its reference.

attrs

a crate::AttrList, can be None

pub fn set_auto_dir(&self, auto_dir: bool)

Sets whether to calculate the bidirectional base direction for the layout according to the contents of the layout; when this flag is on (the default), then paragraphs in self that begin with strong right-to-left characters (Arabic and Hebrew principally), will have right-to-left layout, paragraphs with letters from other scripts will have left-to-right layout. Paragraphs with only neutral characters get their direction from the surrounding paragraphs.

When false, the choice between left-to-right and right-to-left layout is done according to the base direction of the layout’s crate::Context. (See crate::Context::set_base_dir()).

When the auto-computed direction of a paragraph differs from the base direction of the context, the interpretation of crate::Alignment::Left and crate::Alignment::Right are swapped.

auto_dir

if true, compute the bidirectional base direction from the layout’s contents.

pub fn set_ellipsize(&self, ellipsize: EllipsizeMode)

Sets the type of ellipsization being performed for self. Depending on the ellipsization mode ellipsize text is removed from the start, middle, or end of text so they fit within the width and height of layout set with Self::set_width() and Self::set_height().

If the layout contains characters such as newlines that force it to be layed out in multiple paragraphs, then whether each paragraph is ellipsized separately or the entire layout is ellipsized as a whole depends on the set height of the layout. See Self::set_height() for details.

ellipsize

the new ellipsization mode for self

pub fn set_font_description(&self, desc: Option<&FontDescription>)

Sets the default font description for the layout. If no font description is set on the layout, the font description from the layout’s context is used.

desc

the new crate::FontDescription, or None to unset the current font description

pub fn set_height(&self, height: i32)

Sets the height to which the crate::Layout should be ellipsized at. There are two different behaviors, based on whether height is positive or negative.

If height is positive, it will be the maximum height of the layout. Only lines would be shown that would fit, and if there is any text omitted, an ellipsis added. At least one line is included in each paragraph regardless of how small the height value is. A value of zero will render exactly one line for the entire layout.

If height is negative, it will be the (negative of) maximum number of lines per paragraph. That is, the total number of lines shown may well be more than this value if the layout contains multiple paragraphs of text. The default value of -1 means that first line of each paragraph is ellipsized. This behvaior may be changed in the future to act per layout instead of per paragraph. File a bug against pango at http://bugzilla.gnome.org/</ulink> if your code relies on this behavior.

Height setting only has effect if a positive width is set on self and ellipsization mode of self is not crate::EllipsizeMode::None. The behavior is undefined if a height other than -1 is set and ellipsization mode is set to crate::EllipsizeMode::None, and may change in the future.

height

the desired height of the layout in Pango units if positive, or desired number of lines if negative.

pub fn set_indent(&self, indent: i32)

Sets the width in Pango units to indent each paragraph. A negative value of indent will produce a hanging indentation. That is, the first line will have the full width, and subsequent lines will be indented by the absolute value of indent.

The indent setting is ignored if layout alignment is set to crate::Alignment::Center.

indent

the amount by which to indent.

pub fn set_justify(&self, justify: bool)

Sets whether each complete line should be stretched to fill the entire width of the layout. This stretching is typically done by adding whitespace, but for some scripts (such as Arabic), the justification may be done in more complex ways, like extending the characters.

Note that this setting is not implemented and so is ignored in Pango older than 1.18.

justify

whether the lines in the layout should be justified.

pub fn set_line_spacing(&self, factor: f32)

This is supported on crate feature v1_44 only.

Sets a factor for line spacing. Typical values are: 0, 1, 1.5, 2. The default values is 0.

If factor is non-zero, lines are placed so that

baseline2 = baseline1 + factor * height2

where height2 is the line height of the second line (as determined by the font(s)). In this case, the spacing set with Self::set_spacing() is ignored.

If factor is zero, spacing is applied as before.

factor

the new line spacing factor

pub fn set_markup(&self, markup: &str)

Same as Self::set_markup_with_accel(), but the markup text isn’t scanned for accelerators.

markup

marked-up text

length

length of marked-up text in bytes, or -1 if markup is null-terminated

pub fn set_markup_with_accel(&self, markup: &str, accel_marker: char) -> char

Sets the layout text and attribute list from marked-up text (see
markup format``). Replaces the current text and attribute list.
If accel_marker is nonzero, the given character will mark the character following it as an accelerator. For example, accel_marker might be an ampersand or underscore. All characters marked as an accelerator will receive a crate::Underline::Low attribute, and the first character so marked will be returned in accel_char. Two accel_marker characters following each other produce a single literal accel_marker character.

markup

marked-up text (see markup format</link>)

length

length of marked-up text in bytes, or -1 if markup is null-terminated

accel_marker

marker for accelerators in the text

accel_char

return location for first located accelerator, or None

pub fn set_single_paragraph_mode(&self, setting: bool)

If setting is true, do not treat newlines and similar characters as paragraph separators; instead, keep all text in a single paragraph, and display a glyph for paragraph separator characters. Used when you want to allow editing of newlines on a single text line.

setting

new setting

pub fn set_spacing(&self, spacing: i32)

Sets the amount of spacing in Pango unit between the lines of the layout. When placing lines with spacing, Pango arranges things so that

line2.top = line1.bottom + spacing

Note: Since 1.44, Pango defaults to using the line height (as determined by the font) for placing lines. The spacing set with this function is only taken into account when the line-height factor is set to zero with Self::set_line_spacing().

spacing

the amount of spacing

pub fn set_tabs(&self, tabs: Option<&TabArray>)

Sets the tabs to use for self, overriding the default tabs (by default, tabs are every 8 spaces). If tabs is None, the default tabs are reinstated. tabs is copied into the layout; you must free your copy of tabs yourself.

tabs

a crate::TabArray, or None

pub fn set_text(&self, text: &str)

Sets the text of the layout.

This function validates text and renders invalid UTF-8 with a placeholder glyph.

Note that if you have used Self::set_markup() or Self::set_markup_with_accel() on self before, you may want to call Self::set_attributes() to clear the attributes set on the layout from the markup as this function does not clear attributes.

text

the text

length

maximum length of text, in bytes. -1 indicates that the string is nul-terminated and the length should be calculated. The text will also be truncated on encountering a nul-termination even when length is positive.

pub fn set_width(&self, width: i32)

Sets the width to which the lines of the crate::Layout should wrap or ellipsized. The default value is -1: no width set.

width

the desired width in Pango units, or -1 to indicate that no wrapping or ellipsization should be performed.

pub fn set_wrap(&self, wrap: WrapMode)

Sets the wrap mode; the wrap mode only has effect if a width is set on the layout with Self::set_width(). To turn off wrapping, set the width to -1.

wrap

the wrap mode

pub fn xy_to_index(&self, x: i32, y: i32) -> (bool, i32, i32)

Converts from X and Y position within a layout to the byte index to the character at that logical position. If the Y position is not inside the layout, the closest position is chosen (the position will be clamped inside the layout). If the X position is not within the layout, then the start or the end of the line is chosen as described for crate::LayoutLine::x_to_index(). If either the X or Y positions were not inside the layout, then the function returns false; on an exact hit, it returns true.

x

the X offset (in Pango units) from the left edge of the layout.

y

the Y offset (in Pango units) from the top edge of the layout

index_

location to store calculated byte index

trailing

location to store a integer indicating where in the grapheme the user clicked. It will either be zero, or the number of characters in the grapheme. 0 represents the leading edge of the grapheme.

Returns

true if the coordinates were inside text, false otherwise.