Struct pango::TabArray

#[repr(transparent)]
pub struct TabArray { /* private fields */ }

A TabArray contains an array of tab stops.

TabArray can be used to set tab stops in a Layout. Each tab stop has an alignment, a position, and optionally a character to use as decimal point.

Implementations

impl TabArray

pub fn new(initial_size: i32, positions_in_pixels: bool) -> TabArray

Creates an array of @initial_size tab stops.

Tab stops are specified in pixel units if @positions_in_pixels is true, otherwise in Pango units. All stops are initially at position 0.

initial_size

Initial number of tab stops to allocate, can be 0

positions_in_pixels

whether positions are in pixel units

Returns

the newly allocated TabArray, which should be freed with Pango::TabArray::free().

pub fn decimal_point(&self, tab_index: i32) -> char

Available on crate feature v1_50 only.

Gets the Unicode character to use as decimal point.

This is only relevant for tabs with TabAlign::Decimal alignment, which align content at the first occurrence of the decimal point character.

The default value of 0 means that Pango will use the decimal point according to the current locale.

tab_index

the index of a tab stop

pub fn is_positions_in_pixels(&self) -> bool

Returns true if the tab positions are in pixels, false if they are in Pango units.

Returns

whether positions are in pixels.

pub fn size(&mut self) -> i32

Gets the number of tab stops in @self.

Returns

the number of tab stops in the array.

pub fn tab(&self, tab_index: i32) -> (TabAlign, i32)

Gets the alignment and position of a tab stop.

tab_index

tab stop index

Returns

alignment

location to store alignment

location

location to store tab position

pub fn resize(&mut self, new_size: i32)

Resizes a tab array.

You must subsequently initialize any tabs that were added as a result of growing the array.

new_size

new size of the array

pub fn set_decimal_point(&mut self, tab_index: i32, decimal_point: char)

Available on crate feature v1_50 only.

Sets the Unicode character to use as decimal point.

This is only relevant for tabs with TabAlign::Decimal alignment, which align content at the first occurrence of the decimal point character.

By default, Pango uses the decimal point according to the current locale.

tab_index

the index of a tab stop

decimal_point

the decimal point to use

pub fn set_positions_in_pixels(&mut self, positions_in_pixels: bool)

Available on crate feature v1_50 only.

Sets whether positions in this array are specified in pixels.

positions_in_pixels

whether positions are in pixels

pub fn set_tab(&mut self, tab_index: i32, alignment: TabAlign, location: i32)

Sets the alignment and location of a tab stop.

tab_index

the index of a tab stop

alignment

tab alignment

location

tab location in Pango units

pub fn sort(&mut self)

Available on crate feature v1_50 only.

Utility function to ensure that the tab stops are in increasing order.

pub fn to_str(&self) -> GString

Available on crate feature v1_50 only.

Serializes a TabArray to a string.

No guarantees are made about the format of the string, it may change between Pango versions.

The intended use of this function is testing and debugging. The format is not meant as a permanent storage format.

Returns

a newly allocated string

pub fn from_string(text: &str) -> Result<TabArray, BoolError>

Available on crate feature v1_50 only.

Deserializes a TabArray from a string.

This is the counterpart to to_str(). See that functions for details about the format.

text

a string

Returns

a new TabArray